rename 'api' module to 'k8s'
