+++ mkdocs.yml
@@ -16,7 +16,7 @@ nav:
   - Home:
     - Main: index.md
     - Fixtures: fixtures.md
-  - API: mkapi/api/pytest_helm_charts|upper|strict
+  - API: mkapi/k8s/pytest_helm_charts|upper|strict
   - Examples:
     - Basic: 'examples/basic.md'
     - Giant Swarm App Platform: 'examples/giantswarm.md'
