+++ pytest_helm_charts/giantswarm_app_platform/app.py
@@ -7,7 +7,7 @@ import yaml
 from pykube import HTTPClient, ConfigMap
 from pykube.objects import NamespacedAPIObject
 
-from pytest_helm_charts.api.fixtures import NamespaceFactoryFunc
+from pytest_helm_charts.k8s.fixtures import NamespaceFactoryFunc
 from pytest_helm_charts.giantswarm_app_platform.catalog import CatalogFactoryFunc
 from pytest_helm_charts.utils import YamlDict, wait_for_objects_condition, inject_extra
