+++ tests/test_utils.py
@@ -8,7 +8,7 @@ from pykube.objects import NamespacedAPIObject
 from pytest_mock import MockerFixture, MockFixture
 
 from pytest_helm_charts.utils import wait_for_objects_condition
-from pytest_helm_charts.api.job import make_job_object
+from pytest_helm_charts.k8s.job import make_job_object
 
 MockCR = NamespacedAPIObject
